chore: harden permission hook runtime
