+++ package.json
@@ -44,7 +44,7 @@
     "test:all": "npm test && npm run test:performance && npm run test:integration",
     "test:all:coverage": "npm run test:coverage && npm run test:integration:coverage",
     "lint": "eslint --max-warnings=0 --ext .ts src tests scripts",
-    "lint:shell": "shellcheck -x scripts/pretooluse-dollhouse.sh scripts/pretooluse-codex.sh scripts/pretooluse-cursor.sh scripts/pretooluse-gemini.sh scripts/pretooluse-vscode.sh scripts/pretooluse-windsurf.sh scripts/permission-port-discovery.sh",
+    "lint:shell": "shellcheck -x scripts/pretooluse-dollhouse.sh scripts/pretooluse-codex.sh scripts/pretooluse-cursor.sh scripts/pretooluse-gemini.sh scripts/pretooluse-vscode.sh scripts/pretooluse-windsurf.sh scripts/permission-port-discovery.sh scripts/permission-hook-config.sh",
     "lint:fix": "eslint --ext .ts src tests scripts --fix",
     "security:critical": "cross-env \"NODE_OPTIONS=--experimental-vm-modules --no-warnings\" jest --config tests/jest.config.cjs tests/security/tests --testNamePattern=\"(Command Injection|Path Traversal|YAML)\" --maxWorkers=4",
     "security:rapid": "npm run security:critical -- --testTimeout=30000",
@@ -145,6 +145,7 @@
     "dist/seed-elements/**",
     "scripts/pretooluse-dollhouse.sh",
     "scripts/permission-port-discovery.sh",
+    "scripts/permission-hook-config.sh",
     "scripts/pretooluse-vscode.sh",
     "scripts/pretooluse-cursor.sh",
     "scripts/pretooluse-windsurf.sh",
