+++ src/web/routes/permissionRoutes.ts
@@ -17,6 +17,7 @@ import {
   getLastPermissionHookStartupRepairSummary,
   getPermissionHookStatusAsync,
   reconcilePermissionHookStatus,
+  summarizePermissionHookHealth,
 } from '../../utils/permissionHooks.js';
 
 import { SlidingWindowRateLimiter } from '../../utils/SlidingWindowRateLimiter.js';
@@ -498,6 +499,16 @@ export function registerPermissionRoutes(
       const allowOperations = (data.combinedAllowOperations as string[] | undefined) ?? [];
       const confirmOperations = (data.combinedConfirmOperations as string[] | undefined) ?? [];
 
+      const hookStartupRepair = getLastPermissionHookStartupRepairSummary();
+      const hookHealthHost = hookStatus.host ?? hookHost ?? 'managed-host';
+      const hookHealth = summarizePermissionHookHealth({
+        installedHosts: hookStatus.installed || hookStatus.assetsPrepared ? [hookHealthHost] : [],
+        currentHosts: hookStatus.assetsCurrent ? [hookHealthHost] : [],
+        repairedHosts: hookStatus.autoRepaired ? [hookHealthHost] : [],
+        needsRepairHosts: hookStatus.needsRepair ? [hookHealthHost] : [],
+        lastStartupRepair: hookStartupRepair,
+      });
+
       res.json({
         ...(sessionId ? { sessionId } : {}),
         activeElementCount: data.activeElementCount,
@@ -521,7 +532,8 @@ export function registerPermissionRoutes(
         hookAutoRepaired: hookStatus.autoRepaired,
         hookNeedsRepair: hookStatus.needsRepair,
         hookRepairError: hookStatus.repairError,
-        hookStartupRepair: getLastPermissionHookStartupRepairSummary(),
+        hookHealth,
+        hookStartupRepair,
         authority: authorityState,
         authoritySupportedHosts: installedAuthorityHosts,
         authoritySupportedModes: [...PERMISSION_AUTHORITY_MODES],
