+++ tests/unit/web/setupRoutes.test.ts
@@ -2125,6 +2125,7 @@ describe('Setup Tab — Package Inclusion', () => {
     const files = packageJson.files as string[];
     expect(files).toContain('scripts/pretooluse-dollhouse.sh');
     expect(files).toContain('scripts/permission-port-discovery.sh');
+    expect(files).toContain('scripts/permission-hook-config.sh');
     expect(files).toContain('scripts/pretooluse-vscode.sh');
     expect(files).toContain('scripts/pretooluse-cursor.sh');
     expect(files).toContain('scripts/pretooluse-windsurf.sh');
